requirements.txt
fastapi
uvicorn[standard]
scikit-learn==1.6.1
joblib
numpy
pandas
python-dotenv
pydantic